An arena entry is a block index, and a column whose order did not move writes none: +10.2% on the GPU; 1.6.1
Two changes to what the descriptor arena holds. Neither touches the
descriptor word's width, the descriptor count, or the number of radix
passes, so neither costs anything elsewhere in the kernel.

An arena entry is now a 16-bit block index rather than a 32-bit position.
Every position one descriptor emits sits at the same column within its
256-byte block, so the shared eight bits moved into the descriptor's
fourth key byte -- free since 1.5.8 dropped the fourth radix byte -- and
scatter rebuilds (block << 8) | column. +3.2%.

And a column whose order has not moved now writes nothing at all. A
constant column prepends the same byte to every suffix in the run, so the
order holds; col_same already knew which columns those are and used it to
skip the gather and the sort, but not the store. Since a column writes
exactly ord[0..len) into one contiguous slot in order -- the key groups
partition [0, len) in order, so where they fall does not matter -- and a
slot is never written twice, a column can point its descriptors at an
earlier slot and skip the store. DESC_ARENA_REUSE=0 restores the old
behaviour byte for byte.

Three interleaved --bench --gpu=0 pairs at 336 blocks on an RTX 5080:
145.59/145.57/145.64 KH/s against 160.61/160.24/160.29, +10.2% with no
overlap. The profiler puts it in two places: the column walk falls 922.2M
-> 774.9M cycles because a store loop left its inner loop, and expand to
sa falls 340.0M -> 274.9M untouched, because it now reads back only the
slots that were written. The radix sort is flat at 407M, which is the
check that nothing about the descriptors changed.

desc_test correct on all 512 suffix arrays, hash_parallel_test reports
all 512 whole hashes bit-identical to the CPU, go test ./cmd/derostorm/
green, and the miner's start-up check reports matches the CPU exactly.

+++ gpu/derostorm_gpu.cu
@@ -283,7 +283,7 @@ struct Pool {
         DescScratch s;
         s.words  = keys + k + (size_t)ASTRO_MAX_TEXT * 0;
         s.words2 = keys + k + (size_t)ASTRO_MAX_TEXT * 1;
-        s.arena  = (uint32_t*)(words + w + (size_t)ASTRO_MAX_TEXT * 0);
+        s.arena  = (uint16_t*)(words + w + (size_t)ASTRO_MAX_TEXT * 0);
         s.offs   = words + w + (size_t)ASTRO_MAX_TEXT * 1;
         s.mbuf   = words + w + (size_t)ASTRO_MAX_TEXT * 2;
         return s;
