+++ gpu/hash_parallel_test.cu
@@ -62,7 +62,7 @@ struct Pool {
         DescScratch s;
         s.words  = keys + k + (size_t)stride * 0;
         s.words2 = keys + k + (size_t)stride * 1;
-        s.arena  = (uint32_t*)(words + w + (size_t)stride * 0);
+        s.arena  = (uint16_t*)(words + w + (size_t)stride * 0);
         s.offs   = words + w + (size_t)stride * 1;
         s.mbuf   = words + w + (size_t)stride * 2;
         return s;
